# swagger.yaml
openapi: 3.0.3
info:
  title: User Manager API
  description: API REST para gerenciamento de usuários. Permite criar, buscar, atualizar e deletar usuários.
  version: 1.0.0
  contact:
    name: User Manager API
    email: contact@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8080
    description: Servidor de Desenvolvimento

tags:
  - name: Users
    description: Operações relacionadas ao gerenciamento de usuários

paths:
  /api/v1/user:
    get:
      tags:
        - Users
      summary: Listar todos os usuários
      description: Retorna uma lista com todos os usuários cadastrados no sistema
      operationId: findAllUsers
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserData'
              example:
                - id: 1
                  name: "João Silva"
                  email: "joao.silva@example.com"
                  createAt: "2024-01-15T10:30:00"
                - id: 2
                  name: "Maria Santos"
                  email: "maria.santos@example.com"
                  createAt: "2024-01-16T14:20:00"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Users
      summary: Criar um novo usuário
      description: Cria um novo usuário no sistema com os dados fornecidos
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              example1:
                summary: Exemplo de criação de usuário
                value:
                  name: "João Silva"
                  email: "joao.silva@example.com"
      responses:
        '200':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserData'
              example:
                id: 1
                name: "João Silva"
                email: "joao.silva@example.com"
                createAt: "2024-01-15T10:30:00"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/user/{id}:
    get:
      tags:
        - Users
      summary: Buscar usuário por ID
      description: Retorna os dados de um usuário específico através do seu ID
      operationId: findUserById
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário a ser buscado
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Usuário encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserData'
              example:
                id: 1
                name: "João Silva"
                email: "joao.silva@example.com"
                createAt: "2024-01-15T10:30:00"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Users
      summary: Atualizar usuário
      description: Atualiza os dados de um usuário existente através do seu ID
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário a ser atualizado
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            examples:
              example1:
                summary: Exemplo de atualização de usuário
                value:
                  name: "João Silva Atualizado"
                  email: "joao.silva.novo@example.com"
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserData'
              example:
                id: 1
                name: "João Silva Atualizado"
                email: "joao.silva.novo@example.com"
                createAt: "2024-01-15T10:30:00"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Users
      summary: Deletar usuário
      description: Remove um usuário do sistema através do seu ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário a ser deletado
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Usuário deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    UserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Nome completo do usuário
          example: "João Silva"
          minLength: 1
        email:
          type: string
          format: email
          description: Endereço de email do usuário
          example: "joao.silva@example.com"
          minLength: 1
      example:
        name: "João Silva"
        email: "joao.silva@example.com"

    UserData:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único do usuário
          example: 1
        name:
          type: string
          description: Nome completo do usuário
          example: "João Silva"
        email:
          type: string
          format: email
          description: Endereço de email do usuário
          example: "joao.silva@example.com"
        createAt:
          type: string
          format: date-time
          description: Data e hora de criação do usuário
          example: "2024-01-15T10:30:00"
      required:
        - id
        - name
        - email
        - createAt

    Error:
      type: object
      properties:
        message:
          type: string
          description: Mensagem de erro descritiva
          example: "Nome deve ser obrigatório"
        timestamp:
          type: string
          format: date-time
          description: Data e hora em que o erro ocorreu
          example: "2024-01-15T10:30:00"

  responses:
    BadRequest:
      description: Requisição inválida. Os dados fornecidos não atendem aos requisitos de validação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Nome deve ser obrigatório"
            timestamp: "2024-01-15T10:30:00"

    NotFound:
      description: Usuário não encontrado com o ID fornecido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Usuário não encontrado"
            timestamp: "2024-01-15T10:30:00"

    InternalServerError:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Erro interno do servidor"
            timestamp: "2024-01-15T10:30:00"